requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pymysql==1.1.0
ollama==0.1.4
pyyaml==6.0.1
anthropic==0.8.1